Added Otsu binarization function from Leptonica
Wrapped the Binarize.c BinarizeOtsuAdaptiveThreshold from the unmanaged
Leptonica DLL
